fix(confirmationdialog): remove callback on deactivate dialog (#188)
* fix(confirmationdialog): remove callback on deactivate

Confirmationdialog is used as a decorator, so Modal component used inside is always mounted for any
wrapped component. Due to this, listeners to Modal component are not removed and are triggered
uneccessorily. One such example is onEnter callback is triggered even when the modal is not in
action. So, removing callback on deactivation.

* revert manuall package bump

* fix(confirmationdialog): default value to null